Module ledger_mob_core::engine

The Engine provides functionality required by hardware wallets.

This handles Event inputs and returns Output responses to the caller, see [apdu][crate::apdu] for APDU protocol / encoding specifications.

Structs

BlockVersion

A block version number that is known to be less or equal to BlockVersion::MAX

Engine

Engine provides hardware-independent support for MobileCoin wallet operations

Function

StateIter

An iterator over the variants of Self

TokenId

Token Id, used to identify different assets on on the blockchain.

TxDigest

Transaction digest, used to keep a running digest of inputs to the transaction engine to ensure sync between the host and hardware wallet.

Enums

Error

super::Engine errors

Event

Engine input events, typically decoded from request [APDUs][crate::apdu]

FogId

Fog identifier for resolving account information

IdentState

Identity challenge state

Output

Engine outputs (in response to events), typically encoded to response [APDUs][ledger_mob_apdu]

RingState

Ring signing states

State

Engine internal state enumeration

SummaryState

Summarizer state enumeration

TransactionEntity

An entity with whom a transaction can interact, and who can be identified by the TxSummary verification process

Constants

RESP_SIZE

Ring response size (2 * RING_SIZE)

RING_SIZE

Maximum ring size

Traits

Driver

Driver trait provides platform support for Engine instances

FogCert

Fog certificate information for FogId types