Enum ledger_mob_core::engine::Event

pub enum Event {

    None,
    GetWalletKeys {
        account_index: u32,
    },
    GetSubaddressKeys {
        account_index: u32,
        subaddress_index: u64,
    },
    GetKeyImage {
        account_index: u32,
        subaddress_index: u64,
        txout_public_key: TxOutPublic,
    },
    GetRandom,
    IdentSign {
        ident_index: u32,
        ident_uri: String<32>,
        challenge: Vec<u8, 64>,
    },
    IdentGet,
    TxInit {
        account_index: u32,
        num_rings: u8,
    },
    TxSignMemo {
        subaddress_index: u64,
        tx_public_key: TxOutPublic,
        receiver_view_public: SubaddressViewPublic,
        kind: [u8; 2],
        payload: [u8; 48],
    },
    TxSetMessage(Vec<u8, 64>),
    TxSummaryInit {
        message: [u8; 32],
        block_version: u32,
        num_outputs: u32,
        num_inputs: u32,
    },
    TxSummaryAddOutput {
        masked_amount: Option<MaskedAmount>,
        target_key: CompressedRistrettoPublic,
        public_key: CompressedRistrettoPublic,
        associated_to_input_rules: bool,
    },
    TxSummaryAddOutputUnblinding {
        unmasked_amount: UnmaskedAmount,
        address: Option<PublicSubaddress>,
        fog_info: Option<(FogId, [u8; 64])>,
        tx_private_key: Option<TxPrivateKey>,
    },
    TxSummaryAddInput {
        pseudo_output_commitment: CompressedCommitment,
        input_rules_digest: Option<[u8; 32]>,
        unmasked_amount: UnmaskedAmount,
    },
    TxSummaryBuild {
        fee: Amount,
        tombstone_block: u64,
    },
    TxRingInit {
        ring_size: u8,
        real_index: u8,
        subaddress_index: u64,
        value: u64,
        token_id: u64,
        onetime_private_key: Option<TxOnetimeKey>,
    },
    TxSetBlinding {
        blinding: Scalar,
        output_blinding: Scalar,
    },
    TxAddTxout(u8, ReducedTxOut),
    TxSign,
    TxGetKeyImage,
    TxGetResponse {
        index: u8,
    },
    TxComplete,
    TxGetInfo,

}

Engine input events, typically decoded from request APDUs

Variants

None

GetWalletKeys

Fields

account_index: u32

Fetch wallet keys

GetSubaddressKeys

Fields

account_index: u32

subaddress_index: u64

Fetch subaddress keys

GetKeyImage

Fields

account_index: u32

subaddress_index: u64

txout_public_key: TxOutPublic

Fetch key image

GetRandom

Fetch random value via RNG

IdentSign

Fields

ident_index: u32

ident_uri: String<32>

challenge: Vec<u8, 64>

Request BIP-0017 derived ed25519 identity

IdentGet

Fetch signed identity

TxInit

Fields

account_index: u32

num_rings: u8

Initialise transaction

TxSignMemo

Fields

subaddress_index: u64

tx_public_key: TxOutPublic

receiver_view_public: SubaddressViewPublic

kind: [u8; 2]

payload: [u8; 48]

Sign transaction memos

TxSetMessage(Vec<u8, 64>)

Set transaction message

TxSummaryInit

Fields

message: [u8; 32]

block_version: u32

num_outputs: u32

num_inputs: u32

Set transaction summary (replaces TxSetMessage where streaming verification is supported)

TxSummaryAddOutput

Fields

masked_amount: Option<MaskedAmount>

target_key: CompressedRistrettoPublic

public_key: CompressedRistrettoPublic

associated_to_input_rules: bool

Add output to TxSummary

TxSummaryAddOutputUnblinding

Fields

unmasked_amount: UnmaskedAmount

address: Option<PublicSubaddress>

fog_info: Option<(FogId, [u8; 64])>

tx_private_key: Option<TxPrivateKey>

Add output unblinding to TxSummary

TxSummaryAddInput

Fields

pseudo_output_commitment: CompressedCommitment

input_rules_digest: Option<[u8; 32]>

unmasked_amount: UnmaskedAmount

Add input to TxSummary

TxSummaryBuild

Fields

fee: Amount

tombstone_block: u64

Build complete TxSummary

TxRingInit

Fields

ring_size: u8

real_index: u8

subaddress_index: u64

value: u64

token_id: u64

onetime_private_key: Option<TxOnetimeKey>

Initialise ring signing

TxSetBlinding

Fields

blinding: Scalar

output_blinding: Scalar

TxAddTxout(u8, ReducedTxOut)

Add TxOut to ring

TxSign

Sign ring

TxGetKeyImage

Fetch key image

TxGetResponse

Fields

index: u8

Index into ring

Fetch responses

TxComplete

Complete transaction

TxGetInfo

Fetch TX info / state

Implementations

impl Event

pub unsafe fn init(p: *mut Self)

In-place initialisation

Safety

This is safe as long as the pointer is valid

pub fn parse(ins: u8, buff: &[u8]) -> Result<Self, ApduError>

Parse an incoming APDU to engine event

pub fn hash(&self) -> Option<[u8; 32]>

Compute a SHA512_256 hash for state-mutating events, used in the construction of the streaming digest.

This calls out to ledger_mob_apdu::digest methods for consistency between events and APDUs.