Struct ledger_mob_core::engine::Engine

pub struct Engine<DRV: Driver, RNG: CryptoRngCore = OsRng> { /* private fields */ }

Engine provides hardware-independent support for MobileCoin wallet operations

Implementations

impl<DRV: Driver> Engine<DRV>

pub const fn new(drv: DRV) -> Self

Create a new transaction engine instance with the provided driver, using the default OsRng

impl<DRV: Driver, RNG: CryptoRngCore> Engine<DRV, RNG>

pub const fn new_with_rng(drv: DRV, rng: RNG) -> Self

Create a new transaction engine instance with the provided driver and rng

pub unsafe fn init(p: *mut Self, drv: DRV, rng: RNG)

Initialise an uninitialised engine instance pointer, another adventure in stack frame reduction TODO: add checks that init and new_with_rng match

Safety

per-field init is okay so long as we init all fields

pub fn update(&mut self, evt: &Event) -> Result<Output, Error>

Handle incoming transaction events

pub fn state(&self) -> State

Fetch current engine state

pub fn get_account(&self, account_index: u32) -> Account

Fetch an [Account] instance for a given wallet index

pub fn get_subaddress( &self, account_index: u32, subaddress_index: u64, fog_id: FogId ) -> OutputAddress

Fetch a Subaddress instance for a given wallet and subaddress index

pub fn is_unlocked(&self) -> bool

Check whether engine is unlocked (ie. key requests and scanning have been approved)

pub fn unlock(&mut self)

Unlock the engine (allowing key requests and scanning)

pub fn lock(&mut self)

Lock the engine (requires approval for key requests and scanning)

pub fn approve(&mut self)

Approve a pending transaction (advances state to State::Ready)

pub fn deny(&mut self)

Deny a pending transaction

pub fn reset(&mut self)

Reset engine state

pub fn progress(&self) -> Option<usize>

Fetch progress for non-interactive states (summary, ring signing)

pub fn message(&self) -> Option<&[u8]>

Fetch message for transactions in progress

pub fn report(&self) -> Option<&TxSummaryUnblindingReport<MAX_RECORDS>>

Return report if available

pub fn address(&self, h: &ShortAddressHash) -> Option<&OutputAddress>

Resolve address if available

pub fn ident(&self) -> Option<&Ident>

Return ident info if available

pub fn ident_approve(&mut self, approve: bool)

Approve or deny a pending identity request, updating the IdentState